Moved back things from arrange.rs
