+++ rust/template/differential_datalog/program/mod.rs
@@ -16,10 +16,7 @@ mod arrange;
 mod timestamp;
 mod update;
 
-pub use arrange::{
-    concatenate_collections, AggFunc, ArrangeFunc, Arrangement, FilterFunc, FilterMapFunc,
-    FlatMapFunc, InspectFunc, JoinFunc, MapFunc, SemijoinFunc, XFormArrangement, XFormCollection,
-};
+pub use arrange::concatenate_collections;
 pub use timestamp::{TSNested, TupleTS, TS, TS16};
 pub use update::Update;
 
@@ -58,8 +55,8 @@ use timely::dataflow::operators::*;
 use timely::dataflow::scopes::*;
 use timely::dataflow::ProbeHandle;
 use timely::logging::TimelyEvent;
-use timely::progress::timestamp::Refines;
-use timely::progress::Timestamp;
+use timely::order::TotalOrder;
+use timely::progress::{timestamp::Refines, Timestamp};
 use timely::worker::Worker;
 
 type ValTrace<S> = DefaultValTrace<DDValue, DDValue, <S as ScopeParent>::Timestamp, Weight, u32>;
@@ -90,6 +87,41 @@ pub type IdxId = usize;
 /// The first element of the tuple identifies relation; the second is the index
 /// of arrangement for the given relation.
 pub type ArrId = (RelId, usize);
+
+/// Function type used to map the content of a relation
+/// (see `XFormCollection::Map`).
+pub type MapFunc = fn(DDValue) -> DDValue;
+
+/// (see `XFormCollection::FlatMap`).
+pub type FlatMapFunc = fn(DDValue) -> Option<Box<dyn Iterator<Item = DDValue>>>;
+
+/// Function type used to filter a relation
+/// (see `XForm*::Filter`).
+pub type FilterFunc = fn(&DDValue) -> bool;
+
+/// Function type used to simultaneously filter and map a relation
+/// (see `XFormCollection::FilterMap`).
+pub type FilterMapFunc = fn(DDValue) -> Option<DDValue>;
+
+/// Function type used to inspect a relation
+/// (see `XFormCollection::InspectFunc`)
+pub type InspectFunc = fn(&DDValue, TupleTS, Weight) -> ();
+
+/// Function type used to arrange a relation into key-value pairs
+/// (see `XFormArrangement::Join`, `XFormArrangement::Antijoin`).
+pub type ArrangeFunc = fn(DDValue) -> Option<(DDValue, DDValue)>;
+
+/// Function type used to assemble the result of a join into a value.
+/// Takes join key and a pair of values from the two joined relation
+/// (see `XFormArrangement::Join`).
+pub type JoinFunc = fn(&DDValue, &DDValue, &DDValue) -> Option<DDValue>;
+
+/// Function type used to assemble the result of a semijoin into a value.
+/// Takes join key and value (see `XFormArrangement::Semijoin`).
+pub type SemijoinFunc = fn(&DDValue, &DDValue, &()) -> Option<DDValue>;
+
+/// Aggregation function: aggregates multiple values into a single value.
+pub type AggFunc = fn(&DDValue, &[(&DDValue, Weight)]) -> Option<DDValue>;
 
 // TODO: add validating constructor for Program:
 // - relation id's are unique
@@ -223,6 +255,223 @@ impl Dep {
     }
 }
 
+/// Transformations, such as maps, flatmaps, filters, joins, etc. are the building blocks of
+/// DDlog rules.
+///
+/// Different kinds of transformations can be applied only to flat collections,
+/// only to arranged collections, or both. We therefore use separate types to represent
+/// collection and arrangement transformations.
+///
+/// Note that differential sometimes allows the same kind of transformation to be applied to both
+/// collections and arrangements; however the former is implemented on top of the latter and incurs
+/// the additional cost of arranging the collection. We only support the arranged version of these
+/// transformations, forcing the user to explicitly arrange the collection if necessary (or, as much
+/// as possible, keep the data arranged throughout the chain of transformations).
+///
+/// `XFormArrangement` - arrangement transformation.
+#[derive(Clone)]
+pub enum XFormArrangement {
+    /// FlatMap arrangement into a collection
+    FlatMap {
+        description: String,
+        fmfun: &'static FlatMapFunc,
+        /// Transformation to apply to resulting collection.
+        /// `None` terminates the chain of transformations.
+        next: Box<Option<XFormCollection>>,
+    },
+    FilterMap {
+        description: String,
+        fmfun: &'static FilterMapFunc,
+        /// Transformation to apply to resulting collection.
+        /// `None` terminates the chain of transformations.
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Aggregate
+    Aggregate {
+        description: String,
+        /// Filter arrangement before grouping
+        ffun: Option<&'static FilterFunc>,
+        /// Aggregation to apply to each group.
+        aggfun: &'static AggFunc,
+        /// Apply transformation to the resulting collection.
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Join
+    Join {
+        description: String,
+        /// Filter arrangement before joining
+        ffun: Option<&'static FilterFunc>,
+        /// Arrangement to join with.
+        arrangement: ArrId,
+        /// Function used to put together ouput value.
+        jfun: &'static JoinFunc,
+        /// Join returns a collection: apply `next` transformation to it.
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Semijoin
+    Semijoin {
+        description: String,
+        /// Filter arrangement before joining
+        ffun: Option<&'static FilterFunc>,
+        /// Arrangement to semijoin with.
+        arrangement: ArrId,
+        /// Function used to put together ouput value.
+        jfun: &'static SemijoinFunc,
+        /// Join returns a collection: apply `next` transformation to it.
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Return a subset of values that correspond to keys not present in `arrangement`.
+    Antijoin {
+        description: String,
+        /// Filter arrangement before joining
+        ffun: Option<&'static FilterFunc>,
+        /// Arrangement to antijoin with
+        arrangement: ArrId,
+        /// Antijoin returns a collection: apply `next` transformation to it.
+        next: Box<Option<XFormCollection>>,
+    },
+}
+
+impl XFormArrangement {
+    pub fn description(&self) -> &str {
+        match self {
+            XFormArrangement::FlatMap { description, .. } => &description,
+            XFormArrangement::FilterMap { description, .. } => &description,
+            XFormArrangement::Aggregate { description, .. } => &description,
+            XFormArrangement::Join { description, .. } => &description,
+            XFormArrangement::Semijoin { description, .. } => &description,
+            XFormArrangement::Antijoin { description, .. } => &description,
+        }
+    }
+
+    pub(super) fn dependencies(&self) -> FnvHashSet<Dep> {
+        match self {
+            XFormArrangement::FlatMap { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormArrangement::FilterMap { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormArrangement::Aggregate { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormArrangement::Join {
+                arrangement, next, ..
+            } => {
+                let mut deps = match **next {
+                    None => FnvHashSet::default(),
+                    Some(ref n) => n.dependencies(),
+                };
+                deps.insert(Dep::Arr(*arrangement));
+                deps
+            }
+            XFormArrangement::Semijoin {
+                arrangement, next, ..
+            } => {
+                let mut deps = match **next {
+                    None => FnvHashSet::default(),
+                    Some(ref n) => n.dependencies(),
+                };
+                deps.insert(Dep::Arr(*arrangement));
+                deps
+            }
+            XFormArrangement::Antijoin {
+                arrangement, next, ..
+            } => {
+                let mut deps = match **next {
+                    None => FnvHashSet::default(),
+                    Some(ref n) => n.dependencies(),
+                };
+                deps.insert(Dep::Arr(*arrangement));
+                deps
+            }
+        }
+    }
+}
+
+/// `XFormCollection` - collection transformation.
+#[derive(Clone)]
+pub enum XFormCollection {
+    /// Arrange the collection, apply `next` transformation to the resulting collection.
+    Arrange {
+        description: String,
+        afun: &'static ArrangeFunc,
+        next: Box<XFormArrangement>,
+    },
+    /// Apply `mfun` to each element in the collection
+    Map {
+        description: String,
+        mfun: &'static MapFunc,
+        next: Box<Option<XFormCollection>>,
+    },
+    /// FlatMap
+    FlatMap {
+        description: String,
+        fmfun: &'static FlatMapFunc,
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Filter collection
+    Filter {
+        description: String,
+        ffun: &'static FilterFunc,
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Map and filter
+    FilterMap {
+        description: String,
+        fmfun: &'static FilterMapFunc,
+        next: Box<Option<XFormCollection>>,
+    },
+    /// Inspector
+    Inspect {
+        description: String,
+        ifun: &'static InspectFunc,
+        next: Box<Option<XFormCollection>>,
+    },
+}
+
+impl XFormCollection {
+    pub fn description(&self) -> &str {
+        match self {
+            XFormCollection::Arrange { description, .. } => &description,
+            XFormCollection::Map { description, .. } => &description,
+            XFormCollection::FlatMap { description, .. } => &description,
+            XFormCollection::Filter { description, .. } => &description,
+            XFormCollection::FilterMap { description, .. } => &description,
+            XFormCollection::Inspect { description, .. } => &description,
+        }
+    }
+
+    pub fn dependencies(&self) -> FnvHashSet<Dep> {
+        match self {
+            XFormCollection::Arrange { next, .. } => next.dependencies(),
+            XFormCollection::Map { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormCollection::FlatMap { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormCollection::Filter { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormCollection::FilterMap { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+            XFormCollection::Inspect { next, .. } => match **next {
+                None => FnvHashSet::default(),
+                Some(ref n) => n.dependencies(),
+            },
+        }
+    }
+}
+
 /// Datalog rule (more precisely, the body of a rule) starts with a collection
 /// or arrangement and applies a chain of transformations to it.
 #[derive(Clone)]
@@ -262,6 +511,108 @@ impl Rule {
                 let mut deps = xform.dependencies();
                 deps.insert(Dep::Arr(*arr));
                 deps
+            }
+        }
+    }
+}
+
+/// Describes arrangement of a relation.
+#[derive(Clone)]
+pub enum Arrangement {
+    /// Arrange into (key,value) pairs
+    Map {
+        /// Arrangement name; does not have to be unique
+        name: String,
+        /// Function used to produce arrangement.
+        afun: &'static ArrangeFunc,
+        /// The arrangement can be queried using `RunningProgram::query_arrangement`
+        /// and `RunningProgram::dump_arrangement`.
+        queryable: bool,
+    },
+    /// Arrange into a set of values
+    Set {
+        /// Arrangement name; does not have to be unique
+        name: String,
+        /// Function used to produce arrangement.
+        fmfun: &'static FilterMapFunc,
+        /// Apply distinct_total() before arranging filtered collection.
+        /// This is necessary if the arrangement is to be used in an antijoin.
+        distinct: bool,
+    },
+}
+
+impl Arrangement {
+    fn name(&self) -> String {
+        match self {
+            Arrangement::Map { name, .. } => name.clone(),
+            Arrangement::Set { name, .. } => name.clone(),
+        }
+    }
+
+    fn queryable(&self) -> bool {
+        match self {
+            Arrangement::Map { queryable, .. } => *queryable,
+            Arrangement::Set { .. } => false,
+        }
+    }
+
+    fn build_arrangement_root<S>(
+        &self,
+        collection: &Collection<S, DDValue, Weight>,
+    ) -> ArrangedCollection<S, TValAgent<S>, TKeyAgent<S>>
+    where
+        S: Scope,
+        Collection<S, DDValue, Weight>: ThresholdTotal<S, DDValue, Weight>,
+        S::Timestamp: Lattice + Ord + TotalOrder,
+    {
+        match self {
+            Arrangement::Map { afun, .. } => {
+                ArrangedCollection::Map(collection.flat_map(*afun).arrange())
+            }
+            Arrangement::Set {
+                fmfun, distinct, ..
+            } => {
+                let filtered = collection.flat_map(*fmfun);
+                if *distinct {
+                    ArrangedCollection::Set(
+                        filtered
+                            .threshold_total(|_, c| if c.is_zero() { 0 } else { 1 })
+                            .map(|k| (k, ()))
+                            .arrange(), /* arrange_by_self() */
+                    )
+                } else {
+                    ArrangedCollection::Set(filtered.map(|k| (k, ())).arrange())
+                }
+            }
+        }
+    }
+
+    fn build_arrangement<S>(
+        &self,
+        collection: &Collection<S, DDValue, Weight>,
+    ) -> ArrangedCollection<S, TValAgent<S>, TKeyAgent<S>>
+    where
+        S: Scope,
+        S::Timestamp: Lattice + Ord,
+    {
+        match self {
+            Arrangement::Map { afun, .. } => {
+                ArrangedCollection::Map(collection.flat_map(*afun).arrange())
+            }
+            Arrangement::Set {
+                fmfun, distinct, ..
+            } => {
+                let filtered = collection.flat_map(*fmfun);
+                if *distinct {
+                    ArrangedCollection::Set(
+                        filtered
+                            .threshold(|_, c| if c.is_zero() { 0 } else { 1 })
+                            .map(|k| (k, ()))
+                            .arrange(),
+                    )
+                } else {
+                    ArrangedCollection::Set(filtered.map(|k| (k, ())).arrange())
+                }
             }
         }
     }
